Improve Database abstraction, some rename and refactor changes
Rename the 'NoteConformable' to 'NoteProtocol' to comply with the Swift API Design Guidelines (https://swift.org/documentation/api-design-guidelines/#protocols-describing-what-is-should-read-as-nouns; https://swift.org/documentation/api-design-guidelines/#name-according-to-roles, section 3). Below we'll use the renamed 'NoteProtocol' type.

Improve 'Database' abstraction. The 'Database' class was intended to abstract from which particular type of the database do we use: Realm, Core Data, or maybe it's our own implementation of the database functionality, etc. The previous version of the 'Database' abstract was still bound on the realization of the model (in our case it was the 'Note' model) which left us with the inbounding problem unresolved.
So the new solution was to create an abstract class out of the 'Database', move the main logic to the class 'DatabaseRealm' which inherits the 'Database' and overrides all of its methods, computed property and the subscript. We've chosen 'Database' as a class type and not a protocol one because if we do the last one we're receiving the error: "Protocol can only be used as a generic constraint because it has Self or associated type requirements".
'Database' is a generic class with the generic constraint called 'Item' which have to conform to 'NoteProtocol' in order to work. This provides full abstraction from the details of the implementation of a model because the protocol defines abstract interface of the model. This requires the developer when inheriting the 'Database' to specify the model type which your desired database will use as a generic constraint. For instance, for the Realm it will be the subclass of the 'Object' class, for the Core Data — the subclass of the 'NSManagedObjectModel'. Both of them will have to conform to the 'NoteProtocol' as well.
If we wanted to change the database type, e.g. from Realm to Core Data, you should just simply create a new class called 'DatabaseCoreData' which will inherit the 'Database', override all of its methods, and specify the subclass of the 'NSManagedObjectModel' which conforms to 'NoteProtocol' as a generic constraint of the superclass. Also you'll have to specify the generic constraint in every place of the code where you use the 'Database' abstraction.
You shouldn't use the 'Database' as an independent programmatic component. It should be subclassed and overridden in order to be integrated in the app's architecture. Otherwise, if you try to call the methods, computed variable or the subscript your code will receive fatalError and crash. We will have no problem replacing it in different parts of our code because of the structure of the 'Database' and 'NoteProtocol'.

Also change the signature of the 'updateNote(text:forNote:)' method of the 'Database' to 'updateNote(_:withText:)' to comply with the Swift API Design Guidelines (https://swift.org/documentation/api-design-guidelines/#omit-first-argument-if-partial-phrase), 'deleteNote(forReference:)' to 'deleteNote(_:)' (https://swift.org/documentation/api-design-guidelines/#omit-needless-words, https://swift.org/documentation/api-design-guidelines/#omit-first-argument-if-partial-phrase), and make some minor code refactor and file tree structure changes.